requirements.txt
chromadb
ollama
flask
beautifulsoup4
requests
apscheduler
pyyaml
feedparser
pandas
pyarrow
datasets
lxml

